Drop horizontal bar graph, looks bad
Add tooltipFormatter prop
